refactor(#73): TaskCardView を Presentation/Components/ に切り出す
#73 で TaskCardView に existingCount prop を追加し、component 単独で
ViewInspector test を可能にするための事前 refactor。#74 で確立した
component 分離 path に倣う。

挙動の差分なし (behavior-preserving)。
